+++ src/IO/AbstractWKBReader.php
@@ -110,8 +110,8 @@ abstract class AbstractWKBReader
         for ($i = 0; $i < $numCurves; $i++) {
             $curve = $this->readGeometry($buffer, $cs->SRID());
 
-            if (! $curve instanceof Curve) {
-                throw new GeometryIOException('Expected Curve, got ' . $curve->geometryType());
+            if (! $curve instanceof LineString && ! $curve instanceof CircularString) {
+                throw new GeometryIOException('Expected LineString|CircularString, got ' . $curve->geometryType());
             }
 
             $curves[] = $curve;